feat: publish cli package to github packages instead of npm
The cli package is now published to github packages instead of npm.
This is to allow for private packages and to keep all packages in one place.
The .releaserc.json file was updated to use the github packages registry.
The .github/workflows/release.yml file was updated to remove the NPM_TOKEN environment variable.
The packages/cli/package.json file was updated to change the package name to @yotpoltd/cadr-cli.